+++ app/src/main/AndroidManifest.xml
@@ -43,13 +43,6 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity
-            android:name=".TaskCreatorActivity"
-            android:exported="true"
-            android:label="@string/title_add_new_task"
-            android:parentActivityName=".MainActivity"
-            android:theme="@style/AppTheme.NoActionBar"
-            android:windowSoftInputMode="adjustResize" />
         <!-- adjustResize will pan the view up when keyboard opens up. -->
 
         <activity
@@ -94,7 +87,7 @@
             android:name=".OnboardingActivity"
             android:exported="true"
             android:theme="@style/AppTheme.NoActionBar" />
-        <activity android:name=".TaskCreatorActivity3"
+        <activity android:name=".TaskCreatorActivity"
             android:exported="true"
             android:parentActivityName=".MainActivity"
             android:theme="@style/AppTheme"/>
@@ -133,8 +126,6 @@
         <meta-data
             android:name="com.google.android.geo.API_KEY"
             android:value="@string/places_api_key" />
-
-        <activity android:name=".TaskCreatorNew" />
 
     </application>
