+++ app/src/main/java/app/tasknearby/yashcreations/com/tasknearby/DetailActivity.java
@@ -21,7 +21,6 @@ import android.widget.Toast;
 import com.google.firebase.analytics.FirebaseAnalytics;
 import com.squareup.picasso.Picasso;
 
-import org.joda.time.DateTimeConstants;
 import org.joda.time.LocalDate;
 import org.joda.time.LocalTime;
