+++ app/src/main/java/app/tasknearby/yashcreations/com/tasknearby/TaskCreatorActivity.java
@@ -8,6 +8,8 @@ import android.content.Intent;
 import android.content.SharedPreferences;
 import android.content.pm.PackageManager;
 import android.database.Cursor;
+import android.graphics.Color;
+import android.graphics.Typeface;
 import android.net.ConnectivityManager;
 import android.net.Uri;
 import android.os.Bundle;
@@ -18,15 +20,18 @@ import android.support.design.widget.Snackbar;
 import android.support.v4.app.ActivityCompat;
 import android.support.v7.app.ActionBar;
 import android.support.v7.app.AppCompatActivity;
-import android.support.v7.widget.Toolbar;
+import android.support.v7.widget.LinearLayoutManager;
+import android.support.v7.widget.RecyclerView;
 import android.text.TextUtils;
 import android.util.Log;
-import android.view.Menu;
-import android.view.MenuItem;
 import android.view.View;
+import android.view.ViewGroup;
 import android.view.ViewStub;
+import android.view.animation.Animation;
+import android.view.animation.AnimationUtils;
 import android.widget.Button;
 import android.widget.EditText;
+import android.widget.FrameLayout;
 import android.widget.ImageView;
 import android.widget.LinearLayout;
 import android.widget.Switch;
@@ -47,6 +52,8 @@ import org.joda.time.LocalTime;
 
 import java.util.ArrayList;
 import java.util.Calendar;
+import java.util.Collections;
+import java.util.List;
 
 import app.tasknearby.yashcreations.com.tasknearby.database.DbConstants;
 import app.tasknearby.yashcreations.com.tasknearby.models.LocationModel;
@@ -56,12 +63,6 @@ import app.tasknearby.yashcreations.com.tasknearby.utils.DistanceUtils;
 import app.tasknearby.yashcreations.com.tasknearby.utils.WeekdayCodeUtils;
 import app.tasknearby.yashcreations.com.tasknearby.utils.firebase.AnalyticsConstants;
 
-/**
- * Creates a new task and also responsible for editing an old one. For editing, we need to use
- * the getEditModeIntent() method to get the starting intent.
- *
- * @author vermayash8
- */
 public class TaskCreatorActivity extends AppCompatActivity implements View.OnClickListener {
 
     private static final String TAG = TaskCreatorActivity.class.getSimpleName();
@@ -88,11 +89,17 @@ public class TaskCreatorActivity extends AppCompatActivity implements View.OnCli
     private TextView startTimeTv, endTimeTv;
     private TextView startDateTv, endDateTv;
     private TextView unitsTv;
-    private ImageView coverImageView;
+    private ImageView taskImageView, arrowAttachmentImage, arrowScheduleImage;
     private Switch alarmSwitch;
     private Switch anytimeSwitch;
     private Switch repeatSwitch;
     private ViewStub weekdaysStub;
+    private LinearLayout selectLocationLayout, selectImageLayout, attachmentTitleLayout,
+            scheduleTitleLayout, timeIntervalLayout, startTimeLayout, endTimeLayout,
+            startDateLayout, endDateLayout, lockLayoutAttachment, lockLayoutSchdule;
+    private FrameLayout scheduleFrameLayout, attachmentFrameLayout;
+    private RecyclerView locationRecyclerView;
+    private Button saveButton, upgradeAttachmentButton, upgradeScheduleButton;
     private WeekdaysDataSource wds;
 
     private FirebaseAnalytics mFirebaseAnalytics;
@@ -114,11 +121,15 @@ public class TaskCreatorActivity extends AppCompatActivity implements View.OnCli
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_task_creator);
+
+        mFirebaseAnalytics = FirebaseAnalytics.getInstance(this);
+        mTaskRepository = new TaskRepository(getApplicationContext());
+
         setActionBar();
         // Find views and set click listeners.
         initializeViews();
-        mFirebaseAnalytics = FirebaseAnalytics.getInstance(this);
-        mTaskRepository = new TaskRepository(getApplicationContext());
+        setRecentPlaces();
+
         // check if activity has been started for editing a task.
         if (getIntent().hasExtra(EXTRA_EDIT_MODE_TASK_ID)) {
             long taskId = getIntent().getLongExtra(EXTRA_EDIT_MODE_TASK_ID, -1);
@@ -126,6 +137,7 @@ public class TaskCreatorActivity extends AppCompatActivity implements View.OnCli
             fillDataForEditing(taskBeingEdited);
             getSupportActionBar().setTitle(getString(R.string.title_edit_task));
         }
+
     }
 
     /**
@@ -136,168 +148,220 @@ public class TaskCreatorActivity extends AppCompatActivity implements View.OnCli
      * @return intent that can be used in startActivity.
      */
     public static Intent getEditModeIntent(Context context, long taskId) {
-        Intent intent = new Intent(context, TaskCreatorActivity3.class);
+        Intent intent = new Intent(context, TaskCreatorActivity.class);
         intent.putExtra(EXTRA_EDIT_MODE_TASK_ID, taskId);
         return intent;
     }
+
 
     /**
      * Sets the toolbar as actionBar and also sets the up button.
      */
     private void setActionBar() {
-        Toolbar toolbar = findViewById(R.id.toolbar);
-        setSupportActionBar(toolbar);
         ActionBar actionBar = getSupportActionBar();
         if (null != actionBar) {
+            actionBar.setElevation(0);
+            actionBar.setTitle("Add Task");
             actionBar.setDisplayHomeAsUpEnabled(true);
             actionBar.setHomeAsUpIndicator(R.drawable.ic_close_black_24dp);
         }
     }
 
-    @Override
-    public void onClick(View v) {
-        switch (v.getId()) {
-            case R.id.fab_image:
-                if (AppUtils.isPremiumUser(this)) {
-                    addTaskImage();
-                } else {
-                    UpgradeActivity.show(this);
-                }
-                break;
-            case R.id.button_saved_places:
-                Intent savedPlacesIntent = new Intent(this, SavedPlacesActivity.class);
-                startActivityForResult(savedPlacesIntent, REQUEST_CODE_LOCATION_SELECTION);
-                break;
-            case R.id.button_place_picker:
-                onPlacePickerRequested();
-                break;
-            case R.id.text_start_time:
-            case R.id.text_end_time:
-                timeSelectionTriggered((TextView) v);
-                break;
-            case R.id.text_start_date:
-            case R.id.text_end_date:
-                dateSelectionTriggered((TextView) v);
-                break;
-            default:
-                break;
-        }
-    }
 
     /**
      * Finds views by id and sets OnClickListener to them.
      */
     private void initializeViews() {
-        // These don't have an OnClickListener.
+
+        // initializing all views
         taskNameInput = findViewById(R.id.edit_text_task_name);
-        locationNameInput = findViewById(R.id.edit_text_location_name);
+        locationNameInput = findViewById(R.id.editText_location_name);
         reminderRangeInput = findViewById(R.id.edit_text_reminder_range);
         noteInput = findViewById(R.id.edit_text_note);
-        coverImageView = findViewById(R.id.image_task_cover);
+        taskImageView = findViewById(R.id.image_selected_image);
         alarmSwitch = findViewById(R.id.switch_alarm);
+        attachmentFrameLayout = findViewById(R.id.frame_layout_attachment);
+        scheduleFrameLayout = findViewById(R.id.frame_layout_schedule);
+        selectLocationLayout = findViewById(R.id.layout_select_location);
+        selectImageLayout = findViewById(R.id.layout_select_image);
+        attachmentTitleLayout = findViewById(R.id.layout_title_attachment);
+        scheduleTitleLayout = findViewById(R.id.layout_title_schedule);
+        arrowAttachmentImage = findViewById(R.id.image_arrow_attachment);
+        arrowScheduleImage = findViewById(R.id.image_arrow_schedule);
+        unitsTv = findViewById(R.id.text_units);
+        anytimeSwitch = findViewById(R.id.switch_anytime);
+        timeIntervalLayout = findViewById(R.id.layout_time_interval);
+        startTimeLayout = findViewById(R.id.layout_time_from);
+        endTimeLayout = findViewById(R.id.layout_time_to);
+        startTimeTv = findViewById(R.id.text_time_from);
+        endTimeTv = findViewById(R.id.text_time_to);
+        startDateLayout = findViewById(R.id.layout_date_from);
+        endDateLayout = findViewById(R.id.layout_date_to);
+        startDateTv = findViewById(R.id.text_date_from);
+        endDateTv = findViewById(R.id.text_date_to);
+        repeatSwitch = findViewById(R.id.switch_repeat);
+        weekdaysStub = findViewById(R.id.viewStub_repeat);
+        locationRecyclerView = findViewById(R.id.recycler_view_location);
+        saveButton = findViewById(R.id.button_save);
+        upgradeAttachmentButton = findViewById(R.id.button_upgrade_attachment);
+        upgradeScheduleButton = findViewById(R.id.button_upgrade_schedule);
+        lockLayoutAttachment = findViewById(R.id.ll_premium_overlay_lock_attachment);
+        lockLayoutSchdule = findViewById(R.id.ll_premium_overlay_lock_schedule);
 
-        // Set the default value to reminder range input.
+        // setting on click listeners
+        attachmentTitleLayout.setOnClickListener(this);
+        scheduleTitleLayout.setOnClickListener(this);
+        selectImageLayout.setOnClickListener(this);
+        startDateLayout.setOnClickListener(this);
+        endDateLayout.setOnClickListener(this);
+        startTimeLayout.setOnClickListener(this);
+        endTimeLayout.setOnClickListener(this);
+        selectLocationLayout.setOnClickListener(this);
+        saveButton.setOnClickListener(this);
+        upgradeScheduleButton.setOnClickListener(this);
+        upgradeAttachmentButton.setOnClickListener(this);
+        lockLayoutAttachment.setOnClickListener(this);
+        lockLayoutSchdule.setOnClickListener(this);
+
+        // setting defaults and other settings
+
+        // setting default distance range
         String defReminderRange = PreferenceManager.getDefaultSharedPreferences(this)
                 .getString(getString(R.string.pref_distance_range_key),
                         getString(R.string.pref_distance_range_default));
         reminderRangeInput.setText(defReminderRange);
 
-        // image choosing FAB and location buttons.
-        findViewById(R.id.fab_image).setOnClickListener(this);
-        findViewById(R.id.button_saved_places).setOnClickListener(this);
-        findViewById(R.id.button_place_picker).setOnClickListener(this);
+        // setting units
+        setReminderRangeUnits();
 
-        // switch to show/hide the time selection layout.
-        anytimeSwitch = findViewById(R.id.switch_time_interval);
-        anytimeSwitch.setOnCheckedChangeListener((buttonView, isChecked) -> {
-            findViewById(R.id.ll_time_adjust).setVisibility((isChecked) ? View.GONE : View.VISIBLE);
-        });
+        // setting anytime switch
+        anytimeSwitch.setOnCheckedChangeListener(((buttonView, isChecked) -> {
+            timeIntervalLayout.setVisibility((isChecked) ? View.GONE : View.VISIBLE);
+        }));
 
-        // time TextViews.
-        startTimeTv = findViewById(R.id.text_start_time);
-        endTimeTv = findViewById(R.id.text_end_time);
-        startTimeTv.setOnClickListener(this);
-        endTimeTv.setOnClickListener(this);
-        // Set the initial times as tag on time textViews.
+        // setting time interval tags with default value
         startTimeTv.setTag(new LocalTime(0, 0));
         endTimeTv.setTag(new LocalTime(23, 59));
 
-        // date textViews.
-        startDateTv = findViewById(R.id.text_start_date);
-        endDateTv = findViewById(R.id.text_end_date);
-        startDateTv.setOnClickListener(this);
-        endDateTv.setOnClickListener(this);
-        // Set initial dates as text on date textViews.
+        // setting date interval tags with default value
         startDateTv.setTag(new LocalDate());
         endDateTv.setTag(null);
 
-        repeatSwitch = findViewById(R.id.switch_repeat);
-        weekdaysStub = findViewById(R.id.weekdays_stub);
+        // setting repeat switch
         repeatSwitch.setOnCheckedChangeListener((buttonView, isChecked) ->
                 weekdaysStub.setVisibility(isChecked ? View.VISIBLE : View.GONE));
+
+        // setting weekday stub
         setupWeekdayBar();
-
-        // Units text view.
-        unitsTv = findViewById(R.id.text_units);
-        setUnitsText();
     }
 
     /**
-     * When we've a task that is being edited, we've to fill it's attributes into the input fields.
-     *
-     * @param task The task that is being edited.
+     * Specifies the action to be taken when a view is clicked.
      */
-    private void fillDataForEditing(final TaskModel task) {
-        taskNameInput.setText(task.getTaskName());
-        // Set location
-        mSelectedLocation = mTaskRepository.getLocationById(task.getLocationId());
-        // Shows the location name and makes it visible.
-        onLocationSelected();
-        hasSelectedLocation = true;
-        // Set reminder range
-        reminderRangeInput.setText(String.valueOf(task.getReminderRange()));
-        // Set note
-        noteInput.setText(task.getNote());
-        // Setup time.
-        boolean anytime = task.getStartTime().equals(new LocalTime(0, 0))
-                && task.getEndTime().equals(new LocalTime(23, 59));
-        anytimeSwitch.setChecked(anytime);
-        startTimeTv.setText(AppUtils.getReadableTime(this, task.getStartTime()));
-        endTimeTv.setText(AppUtils.getReadableTime(this, task.getEndTime()));
-        startTimeTv.setTag(task.getStartTime());
-        endTimeTv.setTag(task.getEndTime());
+    @Override
+    public void onClick(View v) {
 
-        // Set date.
-        startDateTv.setText(AppUtils.getReadableLocalDate(this, task.getStartDate()));
-        endDateTv.setText(AppUtils.getReadableLocalDate(this, task.getEndDate()));
-        startDateTv.setTag(task.getStartDate());
-        endDateTv.setTag(task.getEndDate());
+        switch (v.getId()) {
+            case R.id.layout_title_attachment:
+                if (attachmentFrameLayout.getVisibility() == View.GONE) {
+                    expandAttachment();
+                } else {
+                    collapseAttachment();
+                }
+                break;
+            case R.id.layout_title_schedule:
+                if (scheduleFrameLayout.getVisibility() == View.GONE) {
+                    expandSchedule();
+                } else {
+                    collapseSchedule();
+                }
+                break;
+            case R.id.layout_select_image:
+                addTaskImage();
+                break;
+            case R.id.layout_time_from:
+                timeSelectionTriggered(startTimeTv);
+                break;
+            case R.id.layout_time_to:
+                timeSelectionTriggered(endTimeTv);
+                break;
+            case R.id.layout_date_from:
+                dateSelectionTriggered(startDateTv);
+                break;
+            case R.id.layout_date_to:
+                dateSelectionTriggered(endDateTv);
+                break;
+            case R.id.layout_select_location:
+                onPlacePickerRequested();
+                break;
+            case R.id.button_save:
+                saveTask();
+                break;
+            case R.id.button_upgrade_attachment:
+            case R.id.button_upgrade_schedule:
+            case R.id.ll_premium_overlay_lock_attachment:
+            case R.id.ll_premium_overlay_lock_schedule:
+                mFirebaseAnalytics.logEvent(AnalyticsConstants.PREMIUM_DIALOG_REQUESTED_BY_BUTTON,
+                        new Bundle());
+                UpgradeActivity.show(TaskCreatorActivity.this);
+                break;
 
-        // Repeat options.
-        repeatSwitch.setChecked(task.getRepeatType() == DbConstants.REPEAT_DAILY);
-        if (wds != null) {
-            int repeatCode = task.getRepeatCode();
-            weekdaysStub.setTag(repeatCode);
-            // Get the day indices to repeat.
-            ArrayList<Integer> dayIndices = WeekdayCodeUtils.getDayIndexListToRepeat(repeatCode);
-            for (int day : dayIndices) {
-                wds.setSelectedDays(day - 1);
-            }
         }
-
-        // Alarm switch
-        alarmSwitch.setChecked(task.getIsAlarmSet() != 0);
-        // Cover image
-        if (task.getImageUri() != null) {
-            coverImageView.setImageURI(Uri.parse(task.getImageUri()));
-            coverImageView.setTag(task.getImageUri());
-        }
-        mFirebaseAnalytics.logEvent(AnalyticsConstants.ANALYTICS_EDIT_TASK, new Bundle());
     }
 
     /**
-     * Triggered when FAB is clicked to add image and when permissions are granted. This also
-     * checks and requests if required permissions are not available.
+     * Returns the slideup animation.
+     */
+    private Animation getSlideUpAnimation() {
+        return AnimationUtils.loadAnimation(getApplicationContext(), R.anim.slide_up);
+    }
+
+    /**
+     * Returns the slidedown animation.
+     */
+    private Animation getSlideDownAnimation() {
+        return AnimationUtils.loadAnimation(getApplicationContext(), R.anim.slide_down);
+    }
+
+    /**
+     * Expands the attachment layout
+     */
+    private void expandAttachment() {
+        attachmentFrameLayout.setVisibility(View.VISIBLE);
+        attachmentFrameLayout.startAnimation(getSlideDownAnimation());
+        arrowAttachmentImage.setImageResource(R.drawable.ic_round_keyboard_arrow_up_24px);
+    }
+
+    /**
+     * Collapses attachment layout.
+     */
+    private void collapseAttachment() {
+        attachmentFrameLayout.setVisibility(View.GONE);
+        attachmentFrameLayout.startAnimation(getSlideUpAnimation());
+        arrowAttachmentImage.setImageResource(R.drawable.ic_round_keyboard_arrow_down_24px);
+    }
+
+    /**
+     * Expands schedule layout.
+     */
+    private void expandSchedule() {
+        scheduleFrameLayout.setVisibility(View.VISIBLE);
+        scheduleFrameLayout.startAnimation(getSlideDownAnimation());
+        arrowScheduleImage.setImageResource(R.drawable.ic_round_keyboard_arrow_up_24px);
+    }
+
+    /**
+     * Collapses schedule layout.
+     */
+    private void collapseSchedule() {
+        scheduleFrameLayout.setVisibility(View.GONE);
+        scheduleFrameLayout.startAnimation(getSlideUpAnimation());
+        arrowScheduleImage.setImageResource(R.drawable.ic_round_keyboard_arrow_down_24px);
+    }
+
+    /**
+     * Handles the adding of the task image.This also checks and requests if required permissions
+     * are not available.
      */
     private void addTaskImage() {
         mFirebaseAnalytics.logEvent(AnalyticsConstants.ANALYTICS_ADD_IMAGE, new Bundle());
@@ -307,15 +371,130 @@ public class TaskCreatorActivity extends AppCompatActivity implements View.OnCli
                     .READ_EXTERNAL_STORAGE}, REQUEST_CODE_STORAGE_PERMISSION);
         } else {
             // Permission is available.
-            Intent pickPhoto = new Intent(Intent.ACTION_PICK,
-                    android.provider.MediaStore.Images.Media.EXTERNAL_CONTENT_URI);
+            Intent pickPhoto = new Intent(Intent.ACTION_PICK, android.provider.MediaStore.Images
+                    .Media.EXTERNAL_CONTENT_URI);
             startActivityForResult(pickPhoto, REQUEST_CODE_GALLERY_IMAGE_PICKER);
         }
     }
 
+    @Override
+    public void onRequestPermissionsResult(int requestCode, @NonNull String[] permissions,
+                                           @NonNull int[] grantResults) {
+        switch (requestCode) {
+            case REQUEST_CODE_STORAGE_PERMISSION:
+                if (grantResults.length > 0
+                        && grantResults[0] == PackageManager.PERMISSION_GRANTED) {
+                    addTaskImage();
+                } else {
+                    Toast.makeText(this, R.string.creator_error_image_permission,
+                            Toast.LENGTH_LONG).show();
+                }
+                break;
+        }
+    }
+
+
+    @Override
+    protected void onActivityResult(int requestCode, int resultCode, Intent data) {
+
+        switch (requestCode) {
+            case REQUEST_CODE_GALLERY_IMAGE_PICKER:
+                if (resultCode == RESULT_OK) {
+                    onTaskImageSelected(data);
+                }
+                break;
+            case REQUEST_CODE_PLACE_PICKER:
+                if (resultCode == RESULT_OK) {
+                    onPlacePickerSuccess(data);
+                }
+                break;
+            case REQUEST_CODE_LOCATION_SELECTION:
+                if (resultCode == RESULT_OK) {
+                    onSavedPlacesSuccess(data);
+                }
+                break;
+            default:
+                Log.w(TAG, "Unknown request code in onActivityResult.");
+        }
+        super.onActivityResult(requestCode, resultCode, data);
+    }
+
     /**
-     * Called when user clicks on time display.
+     * Action to take when user selectes the image.
      */
+    private void onTaskImageSelected(Intent data) {
+        if (data.getData() == null) {
+            Toast.makeText(this, R.string.creator_msg_image_selection_failed, Toast.LENGTH_SHORT)
+                    .show();
+            return;
+        }
+
+        Uri selectedImageUri = data.getData();
+        taskImageView.setVisibility(View.VISIBLE);
+        Picasso.with(this)
+                .load(selectedImageUri)
+                .fit()
+                .centerCrop()
+                .into(taskImageView);
+        // We need to generate the image file path from the uri.
+        String[] filePathColumn = {MediaStore.Images.Media.DATA};
+        Cursor cursor = getContentResolver().query(selectedImageUri, filePathColumn,
+                null, null, null);
+        cursor.moveToFirst();
+        String imageFilePath = cursor.getString(cursor.getColumnIndex(filePathColumn[0]));
+        cursor.close();
+        // Set the path as a tag to the imageView for storing in the database
+        // for retrieval later on.
+        taskImageView.setTag(imageFilePath);
+    }
+
+    /**
+     * sets units for reminder range.
+     */
+    private void setReminderRangeUnits() {
+        SharedPreferences defaultPref = PreferenceManager.getDefaultSharedPreferences(this);
+        String unitsPref = defaultPref.getString(getString(R.string.pref_unit_key), getString(R
+                .string.pref_unit_default));
+        if (unitsPref.equals(getString(R.string.pref_unit_metric))) {
+            unitsTv.setText(getString(R.string.unit_metres));
+        } else {
+            unitsTv.setText(getString(R.string.unit_yards));
+        }
+    }
+
+    private void setupWeekdayBar() {
+        // Assumption: No day is selected initially.
+        weekdaysStub.setTag(0);
+         wds = new WeekdaysDataSource(this, R.id.viewStub_repeat)
+                .setFirstDayOfWeek(Calendar.MONDAY)
+                .setUnselectedColorRes(R.color.dark_grey)
+                .start(new WeekdaysDataSource.Callback() {
+                    /**
+                     * Called every time an item is clicked (selected or deselected).
+                     *
+                     * @param weekdaysDataItem calling getCalendarDayId() on this returns the
+                     *                         day's index as in Java Calendar API. Sunday = 1,
+                     *                         Monday = 2....
+                     */
+                    @Override
+                    public void onWeekdaysItemClicked(int i, WeekdaysDataItem weekdaysDataItem) {
+                        int dayCode = WeekdayCodeUtils
+                                .getDayCodeByCalendarDayId(weekdaysDataItem.getCalendarDayId());
+                        int selection = (int) weekdaysStub.getTag();
+                        // Doing an XOR here so that if tapped again, then the day is removed.
+                        selection ^= dayCode;
+                        weekdaysStub.setTag(selection);
+                        Log.d(TAG, "Selected days : " + selection);
+                    }
+
+                    @Override
+                    public void onWeekdaysSelected(int i, ArrayList<WeekdaysDataItem> arrayList) {
+                    }
+                });
+        // Need to explicitly make it GONE in code.
+        weekdaysStub.setVisibility(View.GONE);
+    }
+
     private void timeSelectionTriggered(TextView v) {
         TimePickerDialog timePickerDialog = new TimePickerDialog(this,
                 (view, hourOfDay, minute) -> {
@@ -349,22 +528,6 @@ public class TaskCreatorActivity extends AppCompatActivity implements View.OnCli
         datePickerDialog.show();
     }
 
-    @Override
-    public void onRequestPermissionsResult(int requestCode, @NonNull String[] permissions,
-                                           @NonNull int[] grantResults) {
-        switch (requestCode) {
-            case REQUEST_CODE_STORAGE_PERMISSION:
-                if (grantResults.length > 0
-                        && grantResults[0] == PackageManager.PERMISSION_GRANTED) {
-                    addTaskImage();
-                } else {
-                    Toast.makeText(this, R.string.creator_error_image_permission,
-                            Toast.LENGTH_LONG).show();
-                }
-                break;
-        }
-    }
-
     /**
      * Triggered when the user clicks on the Pick Place button.
      */
@@ -383,35 +546,22 @@ public class TaskCreatorActivity extends AppCompatActivity implements View.OnCli
         }
     }
 
-    @Override
-    protected void onActivityResult(int requestCode, int resultCode, Intent data) {
-        switch (requestCode) {
-            case REQUEST_CODE_PLACE_PICKER:
-                if (resultCode == RESULT_OK) {
-                    onPlacePickerSuccess(data);
-                }
-                break;
-            case REQUEST_CODE_LOCATION_SELECTION:
-                if (resultCode == RESULT_OK) {
-                    onSavedPlacesSuccess(data);
-                } else if (resultCode == SavedPlacesActivity.RESULT_USE_PLACE_PICKER) {
-                    // We're showing a button in saved places activity when no places are being
-                    // shown to allow the users to select pick a place option from there only.
-                    // Didn't want to implement the same place picker calling functionality there,
-                    // so using this way.
-                    onPlacePickerRequested();
-                }
-                break;
-            case REQUEST_CODE_CAMERA_IMAGE:
-            case REQUEST_CODE_GALLERY_IMAGE_PICKER:
-                if (resultCode == RESULT_OK) {
-                    onImageSelected(data);
-                }
-                break;
-            default:
-                Log.w(TAG, "Unknown request code in onActivityResult.");
+    /**
+     * Checks for internet permission. If internet is not connected, it shows a snackbar and
+     * return false.
+     */
+    private boolean isInternetConnected() {
+        ConnectivityManager cm = (ConnectivityManager) getSystemService(Context
+                .CONNECTIVITY_SERVICE);
+        if (cm != null && cm.getActiveNetworkInfo() == null) {
+            // No internet connection present. Show snackbar.
+            Snackbar snackbar = Snackbar.make(findViewById(android.R.id.content), getString(R
+                            .string.creator_no_internet_error),
+                    Snackbar.LENGTH_SHORT);
+            snackbar.show();
+            return false;
         }
-        super.onActivityResult(requestCode, resultCode, data);
+        return true;
     }
 
     /**
@@ -429,6 +579,14 @@ public class TaskCreatorActivity extends AppCompatActivity implements View.OnCli
     }
 
     /**
+     * Sets the selected location's name to the input textView.
+     */
+    private void onLocationSelected() {
+        locationNameInput.setText(mSelectedLocation.getPlaceName());
+        locationNameInput.setVisibility(View.VISIBLE);
+    }
+
+    /**
      * Gets the result from saved places selection activity and sets the location.
      */
     private void onSavedPlacesSuccess(Intent data) {
@@ -437,45 +595,36 @@ public class TaskCreatorActivity extends AppCompatActivity implements View.OnCli
             return;
         }
         long locationId = data.getLongExtra(SavedPlacesActivity.EXTRA_LOCATION_ID, -1);
+        Log.e("shilpi", locationId + "");
         mSelectedLocation = mTaskRepository.getLocationById(locationId);
         hasSelectedLocation = true;
         onLocationSelected();
+        // TODO: check this. Do we need this or not.
+        setRecentPlaces();
     }
 
     /**
-     * Sets the selected location's name to the input textView.
+     * Set the recent places recycler view if there is any recent place present
      */
-    private void onLocationSelected() {
-        locationNameInput.setText(mSelectedLocation.getPlaceName());
-        findViewById(R.id.text_input_location_name).setVisibility(View.VISIBLE);
-    }
-
-    /**
-     * Sets the task image, Triggered when gallery returns a selected task image.
-     */
-    private void onImageSelected(Intent data) {
-        if (data.getData() == null) {
-            Toast.makeText(this, R.string.creator_msg_image_selection_failed, Toast.LENGTH_SHORT)
-                    .show();
+    private void setRecentPlaces() {
+        List<LocationModel> locations = mTaskRepository.getAllLocations();
+        if (locations.size() == 0) {
             return;
         }
-        Uri selectedImageUri = data.getData();
-        // Use Picasso to load image instead of coverImageView.setImageURI(selectedImageUri);
-        Picasso.with(this)
-                .load(selectedImageUri)
-                .fit()
-                .centerCrop()
-                .into(coverImageView);
-        // We need to generate the image file path from the uri.
-        String[] filePathColumn = {MediaStore.Images.Media.DATA};
-        Cursor cursor = getContentResolver().query(selectedImageUri, filePathColumn,
-                null, null, null);
-        cursor.moveToFirst();
-        String imageFilePath = cursor.getString(cursor.getColumnIndex(filePathColumn[0]));
-        cursor.close();
-        // Set the path as a tag to the imageView for storing in the database
-        // for retrieval later on.
-        coverImageView.setTag(imageFilePath);
+
+        locationRecyclerView.setVisibility(View.VISIBLE);
+        // Sort in descending order of use count.
+        Collections.sort(locations, (o1, o2) -> o2.getUseCount() - o1.getUseCount());
+
+        LinearLayoutManager horizontalLayoutManager = new LinearLayoutManager(this,
+                LinearLayoutManager.HORIZONTAL, false);
+        locationRecyclerView.setLayoutManager(horizontalLayoutManager);
+        if (locations.size() > 5) {
+            locationRecyclerView.setAdapter(new RecyclerAdapter(locations.subList(0, 5)));
+        } else {
+            locationRecyclerView.setAdapter(new RecyclerAdapter(locations));
+        }
+
     }
 
     /**
@@ -498,15 +647,12 @@ public class TaskCreatorActivity extends AppCompatActivity implements View.OnCli
         }
         // If reminder range is not valid, no toast has to be shown. In that case, error msg will
         // be empty.
-        if (errorMsg != null && !errorMsg.isEmpty())
+        if (errorMsg != null && !errorMsg.isEmpty()) {
             Toast.makeText(this, errorMsg, Toast.LENGTH_LONG).show();
+        }
         return false;
     }
 
-
-    /**
-     * Validates the input and saves the result to the database.
-     */
     private void saveTask() {
         if (!isInputValid()) {
             return;
@@ -517,7 +663,7 @@ public class TaskCreatorActivity extends AppCompatActivity implements View.OnCli
         int reminderRange = (int) DistanceUtils.getDistanceToSave(this, enteredReminderRange);
         boolean isAlarmEnabled = alarmSwitch.isChecked();
         String imagePath = null;
-        String selectedImagePath = (String) coverImageView.getTag();
+        String selectedImagePath = (String) taskImageView.getTag();
         if (selectedImagePath != null) {
             imagePath = selectedImagePath;
         }
@@ -598,65 +744,20 @@ public class TaskCreatorActivity extends AppCompatActivity implements View.OnCli
         finish();
     }
 
-    private void setupWeekdayBar() {
-        // Assumption: No day is selected initially.
-        weekdaysStub.setTag(0);
-        wds = new WeekdaysDataSource(this, R.id.weekdays_stub)
-                .setFirstDayOfWeek(Calendar.MONDAY)
-                .setUnselectedColorRes(R.color.dark_grey)
-                .start(new WeekdaysDataSource.Callback() {
-                    /**
-                     * Called every time an item is clicked (selected or deselected).
-                     * @param weekdaysDataItem calling getCalendarDayId() on this returns the
-                     * day's index as in Java Calendar API. Sunday = 1, Monday = 2....
-                     */
-                    @Override
-                    public void onWeekdaysItemClicked(int i, WeekdaysDataItem weekdaysDataItem) {
-                        int dayCode = WeekdayCodeUtils
-                                .getDayCodeByCalendarDayId(weekdaysDataItem.getCalendarDayId());
-                        int selection = (int) weekdaysStub.getTag();
-                        // Doing an XOR here so that if tapped again, then the day is removed.
-                        selection ^= dayCode;
-                        weekdaysStub.setTag(selection);
-                        Log.d(TAG, "Selected days : " + selection);
-                    }
-
-                    @Override
-                    public void onWeekdaysSelected(int i, ArrayList<WeekdaysDataItem> arrayList) {
-                    }
-                });
-        // Need to explicitly make it GONE in code.
-        weekdaysStub.setVisibility(View.GONE);
-    }
-
-
-    @Override
-    public boolean onCreateOptionsMenu(Menu menu) {
-        getMenuInflater().inflate(R.menu.menu_task_creator, menu);
-        return super.onCreateOptionsMenu(menu);
-    }
-
-    @Override
-    public boolean onOptionsItemSelected(MenuItem item) {
-        if (item.getItemId() == R.id.action_save) {
-            saveTask();
-        }
-        return super.onOptionsItemSelected(item);
-    }
-
     /**
-     * Sets proper units according to user's settings.
+     * Restarts service.
      */
-    private void setUnitsText() {
+    private void restartService() {
         SharedPreferences defaultPref = PreferenceManager.getDefaultSharedPreferences(this);
-        String unitsPref = defaultPref.getString(getString(R.string.pref_unit_key), getString(R
-                .string.pref_unit_default));
-        if (unitsPref.equals(getString(R.string.pref_unit_metric))) {
-            unitsTv.setText(getString(R.string.unit_metres));
-        } else {
-            unitsTv.setText(getString(R.string.unit_yards));
+        boolean isAppEnabled = defaultPref.getString(getString(R.string.pref_status_key),
+                getString(R.string.pref_status_default)).equals(getString(R.string
+                .pref_status_enabled));
+        if (isAppEnabled) {
+            AppUtils.stopService(this);
+            AppUtils.startService(this);
         }
     }
+
 
     /**
      * Log task creation events.
@@ -676,36 +777,59 @@ public class TaskCreatorActivity extends AppCompatActivity implements View.OnCli
     }
 
     /**
-     * Restarts service.
+     * When we've a task that is being edited, we've to fill it's attributes into the input fields.
+     *
+     * @param task The task that is being edited.
      */
-    private void restartService() {
-        SharedPreferences defaultPref = PreferenceManager.getDefaultSharedPreferences(this);
-        boolean isAppEnabled = defaultPref.getString(getString(R.string.pref_status_key),
-                getString(R.string.pref_status_default)).equals(getString(R.string
-                .pref_status_enabled));
-        if (isAppEnabled) {
-            AppUtils.stopService(this);
-            AppUtils.startService(this);
+    private void fillDataForEditing(final TaskModel task) {
+        taskNameInput.setText(task.getTaskName());
+        // Set location
+        mSelectedLocation = mTaskRepository.getLocationById(task.getLocationId());
+        // Shows the location name and makes it visible.
+        onLocationSelected();
+        hasSelectedLocation = true;
+        // Set reminder range
+        reminderRangeInput.setText(String.valueOf(task.getReminderRange()));
+        // Set note
+        noteInput.setText(task.getNote());
+        // Setup time.
+        boolean anytime = task.getStartTime().equals(new LocalTime(0, 0))
+                && task.getEndTime().equals(new LocalTime(23, 59));
+        anytimeSwitch.setChecked(anytime);
+        startTimeTv.setText(AppUtils.getReadableTime(this, task.getStartTime()));
+        endTimeTv.setText(AppUtils.getReadableTime(this, task.getEndTime()));
+        startTimeTv.setTag(task.getStartTime());
+        endTimeTv.setTag(task.getEndTime());
+
+        // Set date.
+        startDateTv.setText(AppUtils.getReadableLocalDate(this, task.getStartDate()));
+        endDateTv.setText(AppUtils.getReadableLocalDate(this, task.getEndDate()));
+        startDateTv.setTag(task.getStartDate());
+        endDateTv.setTag(task.getEndDate());
+
+        // Repeat options.
+        repeatSwitch.setChecked(task.getRepeatType() == DbConstants.REPEAT_DAILY);
+        if (wds != null) {
+            int repeatCode = task.getRepeatCode();
+            weekdaysStub.setTag(repeatCode);
+            // Get the day indices to repeat.
+            ArrayList<Integer> dayIndices = WeekdayCodeUtils.getDayIndexListToRepeat(repeatCode);
+            for (int day : dayIndices) {
+                wds.setSelectedDays(day - 1);
+            }
         }
+
+        // Alarm switch
+        alarmSwitch.setChecked(task.getIsAlarmSet() != 0);
+        // Cover image
+        if (task.getImageUri() != null) {
+            taskImageView.setVisibility(View.VISIBLE);
+            taskImageView.setImageURI(Uri.parse(task.getImageUri()));
+            taskImageView.setTag(task.getImageUri());
+        }
+        mFirebaseAnalytics.logEvent(AnalyticsConstants.ANALYTICS_EDIT_TASK, new Bundle());
     }
 
-    /**
-     * Checks for internet permission. If internet is not connected, it shows a snackbar and
-     * return false.
-     */
-    private boolean isInternetConnected() {
-        ConnectivityManager cm = (ConnectivityManager) getSystemService(Context
-                .CONNECTIVITY_SERVICE);
-        if (cm != null && cm.getActiveNetworkInfo() == null) {
-            // No internet connection present. Show snackbar.
-            Snackbar snackbar = Snackbar.make(findViewById(android.R.id.content), getString(R
-                            .string.creator_no_internet_error),
-                    Snackbar.LENGTH_SHORT);
-            snackbar.show();
-            return false;
-        }
-        return true;
-    }
 
     @Override
     protected void onStart() {
@@ -716,27 +840,87 @@ public class TaskCreatorActivity extends AppCompatActivity implements View.OnCli
     }
 
     private void setPremiumLock() {
-        // Find views.
-        LinearLayout premiumLockLayout = findViewById(R.id.ll_premium_overlay_lock);
-        Button upgradeButton = findViewById(R.id.button_upgrade);
-        // Set an onClickListener on them to show the dialog.
-        View.OnClickListener premiumLockListener = v -> {
-            mFirebaseAnalytics.logEvent(AnalyticsConstants.PREMIUM_DIALOG_REQUESTED_BY_BUTTON,
-                    new Bundle());
-            UpgradeActivity.show(TaskCreatorActivity.this);
-        };
-        upgradeButton.setOnClickListener(premiumLockListener);
-        premiumLockLayout.setOnClickListener(premiumLockListener);
-        // Adjust the views in the app.
         if (AppUtils.isPremiumUser(this)) {
-            premiumLockLayout.setVisibility(View.GONE);
+            lockLayoutSchdule.setVisibility(View.GONE);
+            lockLayoutAttachment.setVisibility(View.GONE);
             noteInput.setFocusable(true);
             noteInput.setFocusableInTouchMode(true);
         } else {
-            premiumLockLayout.setVisibility(View.VISIBLE);
+            lockLayoutSchdule.setVisibility(View.VISIBLE);
+            lockLayoutAttachment.setVisibility(View.VISIBLE);
+            noteInput.setFocusable(true);
             // The note input can still gain focus by clicking enter button on keyboard,
             // to avoid this, set it as not focusable.
             noteInput.setFocusable(false);
+        }
+    }
+
+
+    class RecyclerAdapter extends RecyclerView.Adapter<RecyclerAdapter.LocationViewHolder> {
+
+        private List<LocationModel> mLocations;
+
+        public RecyclerAdapter(List<LocationModel> locationModels) {
+
+            this.mLocations = locationModels;
+        }
+
+        @NonNull
+        @Override
+        public RecyclerAdapter.LocationViewHolder onCreateViewHolder(@NonNull ViewGroup parent, int
+                viewType) {
+            View v = getLayoutInflater().inflate(R.layout.list_item_location_chip, parent, false);
+            return new LocationViewHolder(v);
+        }
+
+        @Override
+        public void onBindViewHolder(@NonNull RecyclerAdapter.LocationViewHolder holder, int
+                position) {
+            holder.bind(position);
+        }
+
+        @Override
+        public int getItemCount() {
+            return mLocations.size() + 1;
+        }
+
+        public class LocationViewHolder extends RecyclerView.ViewHolder implements View
+                .OnClickListener {
+            TextView textView;
+
+            public LocationViewHolder(View itemView) {
+                super(itemView);
+                textView = itemView.findViewById(R.id.text_view);
+                textView.setOnClickListener(this);
+            }
+
+            public void bind(int position) {
+
+                if (position == mLocations.size()) {
+                    textView.setText("More");
+                    textView.setTextColor(Color.BLUE);
+                    textView.setTypeface(textView.getTypeface(), Typeface.BOLD);
+//                    textView.setBackground(null);
+                } else {
+                    textView.setText(mLocations.get(position).getPlaceName());
+                }
+            }
+
+            @Override
+            public void onClick(View v) {
+                int position = getLayoutPosition();
+                if (v.getId() == R.id.text_view) {
+                    if (position == mLocations.size()) {
+                        Intent savedPlacesIntent = new Intent(TaskCreatorActivity.this,
+                                SavedPlacesActivity.class);
+                        startActivityForResult(savedPlacesIntent, REQUEST_CODE_LOCATION_SELECTION);
+                    } else {
+                        mSelectedLocation = mLocations.get(position);
+                        hasSelectedLocation = true;
+                        onLocationSelected();
+                    }
+                }
+            }
         }
     }
 }
