--- app/src/main/res/drawable/divider_task_creator.xml
@@ -1,5 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<shape xmlns:android="http://schemas.android.com/apk/res/android">
-    <size android:height="1px"/>
-    <solid android:color="#80909090"/>
-</shape>
