--- app/src/main/res/drawable/ic_arrow_down_black_24dp.xml
@@ -1,5 +1,0 @@
-<vector android:height="24dp" android:tint="#616161"
-    android:viewportHeight="24.0" android:viewportWidth="24.0"
-    android:width="24dp" xmlns:android="http://schemas.android.com/apk/res/android">
-    <path android:fillColor="#FF000000" android:pathData="M7.41,7.84L12,12.42l4.59,-4.58L18,9.25l-6,6 -6,-6z"/>
-</vector>
