--- app/src/main/res/drawable/ic_arrow_drop_down_black_12dp.xml
@@ -1,4 +1,0 @@
-<vector android:height="12dp" android:viewportHeight="24.0"
-    android:viewportWidth="24.0" android:width="12dp" xmlns:android="http://schemas.android.com/apk/res/android">
-    <path android:fillColor="#FF000000" android:pathData="M7,10l5,5 5,-5z"/>
-</vector>
