--- app/src/main/res/drawable/ic_arrow_drop_down_black_24dp.xml
@@ -1,5 +1,0 @@
-<vector android:height="24dp" android:tint="#008080"
-    android:viewportHeight="24.0" android:viewportWidth="24.0"
-    android:width="24dp" xmlns:android="http://schemas.android.com/apk/res/android">
-    <path android:fillColor="#FF000000" android:pathData="M7,10l5,5 5,-5z"/>
-</vector>
