--- app/src/main/res/drawable/ic_arrow_up_grey_24dp.xml
@@ -1,5 +1,0 @@
-<vector android:height="24dp" android:tint="#616161"
-    android:viewportHeight="24.0" android:viewportWidth="24.0"
-    android:width="24dp" xmlns:android="http://schemas.android.com/apk/res/android">
-    <path android:fillColor="#FF000000" android:pathData="M7.41,15.41L12,10.83l4.59,4.58L18,14l-6,-6 -6,6z"/>
-</vector>
