--- app/src/main/res/drawable/ic_keyboard_arrow_right_grey_24dp.xml
@@ -1,5 +1,0 @@
-<vector android:height="24dp" android:tint="#616161"
-    android:viewportHeight="24.0" android:viewportWidth="24.0"
-    android:width="24dp" xmlns:android="http://schemas.android.com/apk/res/android">
-    <path android:fillColor="#FF000000" android:pathData="M8.59,16.34l4.58,-4.59 -4.58,-4.59L10,5.75l6,6 -6,6z"/>
-</vector>
