--- app/src/main/res/drawable/ic_repeat_teal_24dp.xml
@@ -1,5 +1,0 @@
-<vector android:height="24dp" android:tint="#00BFA5"
-    android:viewportHeight="24.0" android:viewportWidth="24.0"
-    android:width="24dp" xmlns:android="http://schemas.android.com/apk/res/android">
-    <path android:fillColor="#FF000000" android:pathData="M7,7h10v3l4,-4 -4,-4v3L5,5v6h2L7,7zM17,17L7,17v-3l-4,4 4,4v-3h12v-6h-2v4z"/>
-</vector>
