+++ app/src/main/res/layout/content_detail.xml
@@ -146,7 +146,7 @@
             app:layout_constraintBottom_toBottomOf="@+id/text_time_range"
             app:layout_constraintStart_toStartOf="@+id/guideline_vertical_16dp"
             app:layout_constraintTop_toTopOf="@+id/text_time_range"
-            app:srcCompat="@drawable/ic_clock_grey_700_24dp" />
+            app:srcCompat="@drawable/ic_round_access_time_24px" />
 
         <ImageView
             android:id="@+id/icon_replay"
